+++ examples/src/spooning.rs
@@ -3,7 +3,6 @@ use std::env;
 use tracing::info;
 use tracing_subscriber::filter::LevelFilter;
 use tracing_subscriber::EnvFilter;
-use workspaces::prelude::*;
 use workspaces::{AccountId, Contract, DevNetwork, Worker};
 
 const STATUS_MSG_WASM_FILEPATH: &str = "./examples/res/status_message.wasm";
@@ -51,7 +50,7 @@ async fn deploy_status_contract(
 
     // This will `call` into `set_status` with the message we want to set.
     contract
-        .call(worker, "set_status")
+        .call("set_status")
         .args_json(serde_json::json!({
             "message": msg,
         }))
@@ -107,7 +106,6 @@ async fn main() -> anyhow::Result<()> {
     // Now grab the state to see that it has indeed been patched:
     let status: String = sandbox_contract
         .view(
-            &worker,
             "get_status",
             serde_json::json!({
                 "account_id": testnet_contract_id,
@@ -124,7 +122,6 @@ async fn main() -> anyhow::Result<()> {
     // See that sandbox state was overriden. Grabbing get_status(sandbox_contract_id) should yield Null
     let result: Option<String> = sandbox_contract
         .view(
-            &worker,
             "get_status",
             serde_json::json!({
                 "account_id": sandbox_contract.id(),