+++ examples/src/status_message.rs
@@ -1,5 +1,4 @@
 use serde_json::json;
-use workspaces::prelude::*;
 
 const STATUS_MSG_WASM_FILEPATH: &str = "./examples/res/status_message.wasm";
 
@@ -10,7 +9,7 @@ async fn main() -> anyhow::Result<()> {
     let contract = worker.dev_deploy(&wasm).await?;
 
     let outcome = contract
-        .call(&worker, "set_status")
+        .call("set_status")
         .args_json(json!({
             "message": "hello_world",
         }))
@@ -20,7 +19,6 @@ async fn main() -> anyhow::Result<()> {
 
     let result: String = contract
         .view(
-            &worker,
             "get_status",
             json!({
                 "account_id": contract.id(),