+++ workspaces/src/network/testnet.rs
@@ -11,7 +11,7 @@ use crate::network::{AllowDevAccountCreation, NetworkClient, NetworkInfo, TopLev
 use crate::result::{CallExecution, CallExecutionDetails, ExecutionOutcome, Result};
 use crate::rpc::{client::Client, tool};
 use crate::types::{AccountId, InMemorySigner, SecretKey};
-use crate::{Account, Contract, CryptoHash};
+use crate::{Account, Contract, CryptoHash, Network, Worker};
 
 const RPC_URL: &str = "https://rpc.testnet.near.org";
 const HELPER_URL: &str = "https://helper.testnet.near.org";
@@ -76,6 +76,7 @@ impl AllowDevAccountCreation for Testnet {}
 impl TopLevelAccountCreator for Testnet {
     async fn create_tla(
         &self,
+        worker: Worker<dyn Network>,
         id: AccountId,
         sk: SecretKey,
         // TODO: return Account only, but then you don't get metadata info for it...
@@ -85,7 +86,7 @@ impl TopLevelAccountCreator for Testnet {
         let signer = InMemorySigner::from_secret_key(id.clone(), sk);
 
         Ok(CallExecution {
-            result: Account::new(id, signer),
+            result: Account::new(id, signer, worker),
             details: CallExecutionDetails {
                 // We technically have not burnt any gas ourselves since someone else paid to
                 // create the account for us in testnet when we used the Helper contract.
@@ -109,15 +110,16 @@ impl TopLevelAccountCreator for Testnet {
 
     async fn create_tla_and_deploy(
         &self,
+        worker: Worker<dyn Network>,
         id: AccountId,
         sk: SecretKey,
         wasm: &[u8],
     ) -> Result<CallExecution<Contract>> {
         let signer = InMemorySigner::from_secret_key(id.clone(), sk.clone());
-        let account = self.create_tla(id.clone(), sk).await?;
+        let account = self.create_tla(worker, id.clone(), sk).await?;
         let account = account.into_result()?;
 
-        let outcome = self.client.deploy(&signer, &id, wasm.into()).await?;
+        let outcome = self.client().deploy(&signer, &id, wasm.into()).await?;
 
         Ok(CallExecution {
             result: Contract::account(account),