+++ workspaces/tests/create_account.rs
@@ -1,6 +1,5 @@
 #![recursion_limit = "256"]
 use test_log::test;
-use workspaces::prelude::*;
 
 #[test(tokio::test)]
 async fn test_subaccount_creation() -> anyhow::Result<()> {
@@ -8,7 +7,7 @@ async fn test_subaccount_creation() -> anyhow::Result<()> {
     let account = worker.dev_create_account().await?;
 
     let sub = account
-        .create_subaccount(&worker, "subaccount")
+        .create_subaccount("subaccount")
         .transact()
         .await?
         .into_result()?;