+++ workspaces/tests/cross_contract.rs
@@ -1,8 +1,7 @@
 use near_sdk::json_types::U128;
 use near_units::parse_near;
-use workspaces::prelude::*;
 use workspaces::result::CallExecutionDetails;
-use workspaces::{AccountId, Contract, Network, Worker};
+use workspaces::{AccountId, Contract};
 
 /// The factory contract used in these tests can be found in
 /// [near-sdk/examples/factory-contract](https://github.com/near/near-sdk-rs/tree/master/examples/factory-contract/high-level).
@@ -13,11 +12,10 @@ const FACTORY_CONTRACT: &[u8] =
 async fn cross_contract_create_contract(
     status_id: &AccountId,
     status_amt: &U128,
-    worker: &Worker<impl Network>,
     contract: &Contract,
 ) -> anyhow::Result<CallExecutionDetails> {
     contract
-        .call(worker, "deploy_status_message")
+        .call("deploy_status_message")
         .args_json((status_id.clone(), status_amt))
         .deposit(parse_near!("50 N"))
         .max_gas()
@@ -35,8 +33,7 @@ async fn test_cross_contract_create_contract() -> anyhow::Result<()> {
     // Expect to fail for trying to create a new contract account with too short of a
     // top level account name, such as purely just "status"
     let status_id: AccountId = "status".parse().unwrap();
-    let outcome =
-        cross_contract_create_contract(&status_id, &status_amt, &worker, &contract).await?;
+    let outcome = cross_contract_create_contract(&status_id, &status_amt, &contract).await?;
     let failures = outcome.failures();
     assert!(
         failures.len() == 1,
@@ -47,8 +44,7 @@ async fn test_cross_contract_create_contract() -> anyhow::Result<()> {
     // Expect to succeed after calling into the contract with expected length for a
     // top level account.
     let status_id: AccountId = "status-top-level-account-long-name".parse().unwrap();
-    let outcome =
-        cross_contract_create_contract(&status_id, &status_amt, &worker, &contract).await?;
+    let outcome = cross_contract_create_contract(&status_id, &status_amt, &contract).await?;
     let failures = outcome.failures();
     assert!(
         failures.is_empty(),
@@ -66,11 +62,11 @@ async fn test_cross_contract_calls() -> anyhow::Result<()> {
     let status_amt = U128::from(parse_near!("35 N"));
 
     let status_id: AccountId = "status-top-level-account-long-name".parse().unwrap();
-    cross_contract_create_contract(&status_id, &status_amt, &worker, &contract).await?;
+    cross_contract_create_contract(&status_id, &status_amt, &contract).await?;
 
     let message = "hello world";
     let result = contract
-        .call(&worker, "complex_call")
+        .call("complex_call")
         .args_json((status_id, message))
         .max_gas()
         .transact()