+++ workspaces/tests/deploy_project.rs
@@ -1,7 +1,6 @@
 #![cfg(feature = "unstable")]
 #![recursion_limit = "256"]
 use test_log::test;
-use workspaces::prelude::*;
 
 #[test(tokio::test)]
 async fn test_dev_deploy_project() -> anyhow::Result<()> {
@@ -9,15 +8,15 @@ async fn test_dev_deploy_project() -> anyhow::Result<()> {
     let wasm = workspaces::compile_project("./tests/test-contracts/status-message").await?;
     let contract = worker.dev_deploy(&wasm).await?;
 
-    let _res = contract
-        .call(&worker, "set_status")
+    contract
+        .call("set_status")
         .args_json(("foo",))
         .max_gas()
         .transact()
         .await?;
 
     let res = contract
-        .call(&worker, "get_status")
+        .call("get_status")
         .args_json((contract.id(),))
         .view()
         .await?;