+++ workspaces/tests/optional_args.rs
@@ -1,7 +1,6 @@
 #![recursion_limit = "256"]
 use near_units::parse_near;
 use test_log::test;
-use workspaces::prelude::*;
 use workspaces::{Contract, DevNetwork, Worker};
 
 async fn init(worker: &Worker<impl DevNetwork>) -> anyhow::Result<Contract> {
@@ -10,7 +9,7 @@ async fn init(worker: &Worker<impl DevNetwork>) -> anyhow::Result<Contract> {
         .await?;
 
     contract
-        .call(worker, "new_default_meta")
+        .call("new_default_meta")
         .args_json(serde_json::json!({
             "owner_id": contract.id(),
             "total_supply": parse_near!("1,000,000,000 N").to_string(),
@@ -27,7 +26,7 @@ async fn test_empty_args_error() -> anyhow::Result<()> {
     let contract = init(&worker).await?;
 
     let res = contract
-        .call(&worker, "storage_unregister")
+        .call("storage_unregister")
         .max_gas()
         .deposit(1)
         .transact()
@@ -43,7 +42,7 @@ async fn test_optional_args_present() -> anyhow::Result<()> {
     let contract = init(&worker).await?;
 
     let res = contract
-        .call(&worker, "storage_unregister")
+        .call("storage_unregister")
         .args_json(serde_json::json!({
             "force": true
         }))